12 Image Segmentation/workspace/requirements.txt
argparse
requests
Pillow
torch
torchvision
